+++ pb_plugins/autogen/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+""" Autogenerator for the DroneCore SDK """
+
+
+from .autogen import AutoGen
