+++ pb_plugins/requirements.txt
@@ -1,0 +1,2 @@
+protobuf
+jinja2
