+++ pb_plugins/run_autogen.sh
@@ -1,0 +1,3 @@
+#!/bin/sh -ex
+
+python -m autogen
