+++ pb_plugins/templates/swift/method_completable.j2
