+++ pb_plugins/templates/swift/method_observable.j2
