+++ pb_plugins/templates/swift/method_single.j2
